pyproject.toml
[tool.poetry]
name = "jinjacraft"
version = "1.3.0"
description = "Python command-line tool to generates text files based on Jinja2 template and YAML data file."
authors = ["Steve De Jongh <dejongh.st@gmail.com>"]
readme = "README.md"
repository = "https://github.com/sdejongh/jinjacraft"
license = "MIT"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Text Processing :: General",
    "Topic :: Utilities",
]
keywords = ["jinja2", "yaml", "template", "command-line"]

[tool.poetry.dependencies]
python = ">=3.9"
jinja2 = "^3.1.2"
pyyaml = "^6.0.1"
pytest = "^7.4.0"

[tool.poetry.scripts]
jinjacraft = "jinjacraft.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/sdejongh/jinjacraft/issues"
